抽出 `mods/request`
